style(react-app): rename status.tsx to Status.tsx
